+++ src/migrations/Install.php
@@ -11,11 +11,12 @@
 
 namespace doublesecretagency\googlemaps\migrations;
 
+use Craft;
 use craft\db\Migration;
 
 /**
  * Installation Migration
- * @since 3.0.0
+ * @since 4.0.0
  */
 class Install extends Migration
 {
@@ -25,9 +26,24 @@ class Install extends Migration
      */
     public function safeUp()
     {
-        $this->createTables();
-        $this->createIndexes();
-        $this->addForeignKeys();
+        // Check whether the Smart Map plugin is installed and enabled
+        $pluginInstalled = Craft::$app->plugins->isPluginEnabled('smart-map');
+
+        // Check whether the `smartmap_addresses` table exists
+        $tableExists = $this->db->tableExists('{{%smartmap_addresses}}');
+
+        // If the plugin is installed and the table exists
+        if ($pluginInstalled && $tableExists) {
+
+            // Migrate everything from Smart Map
+            FromSmartMap::update($this);
+
+        } else {
+
+            // Configure the plugin from scratch
+            FromScratch::update($this);
+
+        }
     }
 
     /**
@@ -36,50 +52,6 @@ class Install extends Migration
     public function safeDown()
     {
         $this->dropTableIfExists('{{%googlemaps_addresses}}');
-    }
-
-    /**
-     * Creates the tables.
-     */
-    protected function createTables()
-    {
-        $this->createTable('{{%googlemaps_addresses}}', [
-            'id'          => $this->primaryKey(),
-            'elementId'   => $this->integer()->notNull(),
-            'fieldId'     => $this->integer()->notNull(),
-            'formatted'   => $this->string(),
-            'raw'         => $this->text(),
-            'street1'     => $this->string(),
-            'street2'     => $this->string(),
-            'city'        => $this->string(),
-            'state'       => $this->string(),
-            'zip'         => $this->string(),
-            'country'     => $this->string(),
-            'lat'         => $this->decimal(12, 8),
-            'lng'         => $this->decimal(12, 8),
-            'zoom'        => $this->tinyInteger(2),
-            'dateCreated' => $this->dateTime()->notNull(),
-            'dateUpdated' => $this->dateTime()->notNull(),
-            'uid'         => $this->uid(),
-        ]);
-    }
-
-    /**
-     * Creates the indexes.
-     */
-    protected function createIndexes()
-    {
-        $this->createIndex(null, '{{%googlemaps_addresses}}', ['elementId']);
-        $this->createIndex(null, '{{%googlemaps_addresses}}', ['fieldId']);
-    }
-
-    /**
-     * Adds the foreign keys.
-     */
-    protected function addForeignKeys()
-    {
-        $this->addForeignKey(null, '{{%googlemaps_addresses}}', ['elementId'], '{{%elements}}', ['id'], 'CASCADE');
-        $this->addForeignKey(null, '{{%googlemaps_addresses}}', ['fieldId'],   '{{%fields}}',   ['id'], 'CASCADE');
     }
 
 }