+++ CHANGELOG.md
@@ -1,1 +1,10 @@
 # Changelog
+
+## 4.0.0 - UNRELEASED
+
+### Added
+- Everything.
+
+**This plugin is the direct successor to the Smart Map plugin.** Since the last version of Smart Map ended at 3.x, the Google Maps plugin picks up with version 4.0.
+
+To replace Smart Map, simply install Google Maps. Your data will be migrated automatically.